requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.0.0
openpyxl>=3.1.2
django-filter>=24.2
